requirements.txt
streamlit
Pillow
python-dotenv
pydub
google-generativeai
googletrans
speechrecognition
